+++ src/Internals/Runner.php
@@ -14,6 +14,7 @@ final class Runner {
     use Runner\HasChannels;
     use Runner\HasEater;
     use Runner\HasSharedProgressBar;
+    use Runner\HasSharedConsole;
 
     use Runner\ManagesWorkers;
     use Runner\ManagesTasks;
@@ -33,6 +34,7 @@ final class Runner {
     protected function afterListening(): void {
         $this->stopEater();
         $this->stopRunningTasks();
+        $this->stopConsole();
         $this->closeChannels();
     }
 
@@ -244,6 +246,7 @@ final class Runner {
     }
 
     public function __destruct() {
+        $this->stopConsole();
         $this->stopProgressBar();
     }